requirements.txt
chrisapp~=2.4.0
nose~=1.3.7
pillow==8.2.0
